refactor(domain): ProductDao interface삭제
